+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Changelog
-
-## Unreleased
+## 3.11.0
 
 - Added test for `retried` field on transaction.sale response